src: Code improvements and fixes.
Switching to IDisposable because static shouldn't be used when keeping state. Also, using an array for the state so it can be reliably zeroed, although I expect this worsens performance. Then this fixes implementation mistakes for ARM. I naively thought the ARM API was equivalent to the x86 API in .NET because they're named the same and have the same parameters. I should've used GitHub Actions tests to get access to ARM.